+++ packages/ledger-icrc/src/utils/ledger.utils.ts
@@ -4,10 +4,16 @@ import {
   encodeBase32,
   hexStringToUint8Array,
   isNullish,
+  nonNullish,
   notEmptyString,
   uint8ArrayToHexString,
 } from "@dfinity/utils";
-import type { IcrcAccount } from "../types/ledger.responses";
+import type {
+  IcrcAccount,
+  IcrcTokenMetadata,
+  IcrcTokenMetadataResponse,
+} from "../types/ledger.responses";
+import { IcrcMetadataResponseEntries } from "../types/ledger.responses";
 
 const MAX_SUBACCOUNT_HEX_LENGTH = 64;
 
@@ -88,4 +94,61 @@ export const decodeIcrcAccount = (accountString: string): IcrcAccount => {
   }
 
   return account;
+};
+
+/**
+ * Maps the token metadata information from a ledger response into a structured record.
+ *
+ * This utility processes an array of metadata key-value pairs provided by the ledger
+ * and extracts specific fields, such as symbol, name, fee, decimals, and logo. It then
+ * constructs a `IcrcTokenMetadata` record. If any required fields are missing,
+ * the function returns `undefined`.
+ *
+ * @param {IcrcTokenMetadataResponse} response - An array of key-value pairs representing token metadata.
+ *
+ * @returns {IcrcTokenMetadata | undefined} - A structured metadata record or `undefined` if required fields are missing.
+ */
+export const mapTokenMetadata = (
+  response: IcrcTokenMetadataResponse,
+): IcrcTokenMetadata | undefined => {
+  const nullishToken = response.reduce<Partial<IcrcTokenMetadata>>(
+    (acc, [key, value]) => {
+      switch (key) {
+        case IcrcMetadataResponseEntries.SYMBOL:
+          acc = { ...acc, ...("Text" in value && { symbol: value.Text }) };
+          break;
+        case IcrcMetadataResponseEntries.NAME:
+          acc = { ...acc, ...("Text" in value && { name: value.Text }) };
+          break;
+        case IcrcMetadataResponseEntries.FEE:
+          acc = { ...acc, ...("Nat" in value && { fee: value.Nat }) };
+          break;
+        case IcrcMetadataResponseEntries.DECIMALS:
+          acc = {
+            ...acc,
+            ...("Nat" in value && { decimals: Number(value.Nat) }),
+          };
+          break;
+        case IcrcMetadataResponseEntries.LOGO:
+          acc = { ...acc, ...("Text" in value && { icon: value.Text }) };
+      }
+
+      return acc;
+    },
+    {},
+  );
+
+  const isIcrcTokenMetadata = (
+    arg: Partial<IcrcTokenMetadata>,
+  ): arg is IcrcTokenMetadata =>
+    nonNullish(arg.symbol) &&
+    nonNullish(arg.name) &&
+    nonNullish(arg.fee) &&
+    nonNullish(arg.decimals);
+
+  if (!isIcrcTokenMetadata(nullishToken)) {
+    return undefined;
+  }
+
+  return nullishToken;
 };